diminui volume do Sound e consertando warnings

+++ src/GameObject.cpp
@@ -14,12 +14,12 @@ GameObject::~GameObject(){
 }
 
 void GameObject::Update(float dt){
-	for (int i = 0; i < components.size(); i++)
+	for (int i = 0; i < (int) components.size(); i++)
 		components[i]->Update(dt);
 }
 
 void GameObject::Render(){
-	for (int i = 0; i < components.size(); i++)
+	for (int i = 0; i < (int) components.size(); i++)
 		components[i]->Render();
 }
 
@@ -28,7 +28,7 @@ void GameObject::AddComponent(Component* cpt){
 }
 
 void GameObject::RemoveComponent(Component* cpt){
-	for (int i = 0; i < components.size(); i++){
+	for (int i = 0; i < (int) components.size(); i++){
 		if (components[i].get() == cpt){
 			components.erase(components.begin()+i);
 			return;
@@ -38,7 +38,7 @@ void GameObject::RemoveComponent(Component* cpt){
 
 Component* GameObject::GetComponent(std::string type){
 	Component* result = nullptr;
-	for (int i = 0; i < components.size(); i++){
+	for (int i = 0; i < (int) components.size(); i++){
 		if (components[i]->Is(type)){
 			result = components[i].get();
 			break;
